api/requirements.txt
flask
pillow
redis
uwsgi
